prototype(HB): exchange entry progress
