+++ Blockchain/Homebrew/Exchange/Interactors/ExchangeCreateInteractor.swift
@@ -99,7 +99,7 @@ extension ExchangeCreateInteractor: ExchangeCreateInput {
             Logger.shared.error("Updating input with no model")
             return
         }
-        model.volume = Decimal(string: inputs.activeInput.input)!
+//        model.volume = Decimal(string: inputs.activeInput.input)!
 
         // Update interface to reflect what has been typed
         updateOutput()
